add huge reorg to get new tabular interfaced weights bootstrapped into the library
