+++ pysal/common.py
@@ -17,6 +17,7 @@ except:
     raise
 
 RTOL = .00001
+ATOL = 1e-7
 
 import copy
 import math
