+++ pysal/esda/smoothing.py
@@ -13,13 +13,13 @@ Author(s):
 
 __author__ = "Myunghwa Hwang <mhwang4@gmail.com>, David Folch <dfolch@asu.edu>, Luc Anselin <luc.anselin@asu.edu>, Serge Rey <srey@asu.edu"
 
-import pysal
-from pysal.weights import comb, Kernel
-from pysal.cg import Point, Ray, LineSegment
-from pysal.cg import get_angle_between, get_points_dist, get_segment_point_dist
-from pysal.cg import get_point_at_angle_and_dist, convex_hull
-from pysal.common import np, KDTree
-from pysal.weights.spatial_lag import lag_spatial as slag
+from ..weights.util import comb
+from ..weights import Kernel
+from ..cg import Point, Ray, LineSegment
+from ..cg import get_angle_between, get_points_dist, get_segment_point_dist
+from ..cg import get_point_at_angle_and_dist, convex_hull
+from ..common import np, KDTree
+from ..weights.spatial_lag import lag_spatial as slag
 from scipy.stats import gamma, norm, chi2, poisson
 
 __all__ = ['Excess_Risk', 'Empirical_Bayes', 'Spatial_Empirical_Bayes', 'Spatial_Rate', 'Kernel_Smoother', 'Age_Adjusted_Smoother', 'Disk_Smoother', 'Spatial_Median_Rate', 'Spatial_Filtering', 'Headbanging_Triples', 'Headbanging_Median_Rate', 'flatten', 'weighted_median', 'sum_by_n', 'crude_age_standardization', 'direct_age_standardization', 'indirect_age_standardization', 'standardized_mortality_ratio', 'choynowski', 'assuncao_rate']
