+++ pysal/weights/Distance.py
@@ -1,28 +1,20 @@
-"""
-Distance based spatial weights.
-
-"""
-
-__author__ = "Sergio J. Rey <srey@asu.edu> "
-
 import pysal
-import scipy.spatial
-from pysal.common import KDTree
-from pysal.weights import W
-import scipy.stats
+from ..common import KDTree
+from .weights import W
 import numpy as np
 from util import isKDTree
+from warnings import warn as Warn
+from .util import get_ids, get_points_array_from_shapefile, get_points_array
 
 __all__ = ["knnW", "Kernel", "DistanceBand"]
 
-def knnW(data, k=2, p=2, ids=None):
+class KNN(W):
     """
     Creates nearest neighbor weights matrix based on k nearest
     neighbors.
 
     Parameters
     ----------
-
     kdtree      : object
                   PySAL KDTree or ArcKDTree where KDtree.data is array (n,k)
                   n observations on k characteristics used to measure
@@ -47,16 +39,15 @@ def knnW(data, k=2, p=2, ids=None):
 
     Examples
     --------
-
     >>> points = [(10, 10), (20, 10), (40, 10), (15, 20), (30, 20), (30, 30)]
     >>> kd = pysal.cg.kdtree.KDTree(np.array(points))
-    >>> wnn2 = pysal.knnW(kd, 2)
+    >>> wnn2 = pysal.KNN(kd, 2)
     >>> [1,3] == wnn2.neighbors[0]
     True
 
     ids
 
-    >>> wnn2 = knnW(kd,2)
+    >>> wnn2 = KNN(kd,2)
     >>> wnn2[0]
     {1: 1.0, 3: 1.0}
     >>> wnn2[1]
@@ -64,7 +55,7 @@ def knnW(data, k=2, p=2, ids=None):
 
     now with 1 rather than 0 offset
 
-    >>> wnn2 = knnW(kd, 2, ids=range(1,7))
+    >>> wnn2 = KNN(kd, 2, ids=range(1,7))
     >>> wnn2[1]
     {2: 1.0, 4: 1.0}
     >>> wnn2[2]
@@ -80,28 +71,228 @@ def knnW(data, k=2, p=2, ids=None):
     See Also
     --------
     pysal.weights.W
-
     """
-    if isKDTree(data):
-        kdt = data
-        data = kdt.data
-    else:
-        kdt = KDTree(data)
-    nnq = kdt.query(data, k=k+1, p=p)
-    info = nnq[1]
-
-    neighbors = {}
-    for i, row in enumerate(info):
-        row = row.tolist()
-        if i in row:
+    def __init__(self, data, k=2, p=2, ids=None, radius=None, distance_metric='euclidean'):
+        if isKDTree(data):
+            self.kdtree = data
+            self.data = data.data
+        else:
+            self.data = data
+            self.kdtree = KDTree(data, radius=radius, distance_metric=distance_metric)
+        self.k = k 
+        self.p = p
+        this_nnq = self.kdtree.query(self.data, k=k+1, p=p)
+        
+        to_weight = this_nnq[1]
+        if ids is None:
+            ids = list(range(to_weight.shape[0]))
+        
+        neighbors = {}
+        for i,row in enumerate(to_weight):
+            row = row.tolist()
             row.remove(i)
-            focal = i
-        if ids:
-            row = [ ids[j] for j in row]
+            row = [ids[j] for j in row]
             focal = ids[i]
-        neighbors[focal] = row
-    return pysal.weights.W(neighbors,  id_order=ids)
+            neighbors[focal] = row
+        W.__init__(self, neighbors, id_order=ids)
+    
+    @classmethod
+    def from_shapefile(cls, filepath, **kwargs):
+        """
+        Nearest neighbor weights from a shapefile.
 
+        Parameters
+        ----------
+
+        data       : string
+                     shapefile containing attribute data.
+        k          : int
+                     number of nearest neighbors
+        p          : float
+                     Minkowski p-norm distance metric parameter:
+                     1<=p<=infinity
+                     2: Euclidean distance
+                     1: Manhattan distance
+        ids        : list
+                     identifiers to attach to each observation
+        radius     : float
+                     If supplied arc_distances will be calculated
+                     based on the given radius. p will be ignored.
+
+        Returns
+        -------
+
+        w         : KNN
+                    instance; Weights object with binary weights.
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> x,y=np.indices((5,5))
+        >>> x.shape=(25,1)
+        >>> y.shape=(25,1)
+        >>> data=np.hstack([x,y])
+        >>> wnn2=knnW_from_array(data,k=2)
+        >>> wnn4=knnW_from_array(data,k=4)
+        >>> set([1, 5, 6, 2]) == set(wnn4.neighbors[0])
+        True
+        >>> set([0, 1, 10, 6]) == set(wnn4.neighbors[5])
+        True
+        >>> set([1, 5]) == set(wnn2.neighbors[0])
+        True
+        >>> set([0,6]) == set(wnn2.neighbors[5])
+        True
+        >>> "%.2f"%wnn2.pct_nonzero
+        '8.00'
+        >>> wnn4.pct_nonzero
+        16.0
+        >>> wnn4=knnW_from_array(data,k=4)
+        >>> set([ 1,5,6,2]) == set(wnn4.neighbors[0])
+        True
+
+        Notes
+        -----
+
+        Ties between neighbors of equal distance are arbitrarily broken.
+
+        See Also
+        --------
+        :class:`pysal.weights.W`
+        """
+        return cls(get_points_array_from_shapefile(filepath), **kwargs)
+    
+    @classmethod
+    def from_array(cls, array, **kwargs):
+        """
+        Creates nearest neighbor weights matrix based on k nearest
+        neighbors.
+
+        Parameters
+        ----------
+        array       : np.ndarray
+                      (n, k) array representing n observations on 
+                      k characteristics used to measure distances 
+                      between the n objects
+        **kwargs    : keyword arguments, see Rook
+
+        Returns
+        -------
+        w         : W
+                    instance
+                    Weights object with binary weights
+
+        Examples
+        --------
+        >>> points = [(10, 10), (20, 10), (40, 10), (15, 20), (30, 20), (30, 30)]
+        >>> wnn2 = pysal.KNN.from_array(points, 2)
+        >>> [1,3] == wnn2.neighbors[0]
+        True
+
+        ids
+
+        >>> wnn2 = KNN.from_array(points,2)
+        >>> wnn2[0]
+        {1: 1.0, 3: 1.0}
+        >>> wnn2[1]
+        {0: 1.0, 3: 1.0}
+
+        now with 1 rather than 0 offset
+
+        >>> wnn2 = KNN.from_array(points, 2, ids=range(1,7))
+        >>> wnn2[1]
+        {2: 1.0, 4: 1.0}
+        >>> wnn2[2]
+        {1: 1.0, 4: 1.0}
+        >>> 0 in wnn2.neighbors
+        False
+
+        Notes
+        -----
+
+        Ties between neighbors of equal distance are arbitrarily broken.
+
+        See Also
+        --------
+        KNN
+        """
+        return cls(array, **kwargs)
+
+    @classmethod
+    def from_dataframe(cls, df, geom_col='geometry', ids=None, **kwargs):
+        """
+        Make KNN weights from a dataframe.
+
+        Parameters
+        ----------
+        df      :   pandas.dataframe
+                    a dataframe with a geometry column that can be used to
+                    construct a W object
+        geom_col :   string
+                    column name of the geometry stored in df
+        ids     :   string or iterable
+                    if string, the column name of the indices from the dataframe
+                    if iterable, a list of ids to use for the W
+                    if None, df.index is used.
+
+        See Also
+        --------
+        KNN
+        """
+        pts = get_points_array(df[geom_col])
+        if ids is None:
+            ids = df.index.tolist()
+        elif isinstance(ids, str):
+            ids = df[ids].tolist()
+        return cls(pts, ids=ids, **kwargs)
+
+    def reweight(self, k=None, p=None, new_data=None, new_ids=None, inplace=True):
+        """
+        Redo K-Nearest Neighbor weights construction using given parameters
+
+        Parameters
+        ----------
+        new_data    : np.ndarray
+                      an array containing additional data to use in the KNN
+                      weight
+        new_ids     : list
+                      a list aligned with new_data that provides the ids for
+                      each new observation
+        inplace     : bool
+                      a flag denoting whether to modify the KNN object 
+                      in place or to return a new KNN object
+        k           : int
+                      number of nearest neighbors
+        p           : float
+                      Minkowski p-norm distance metric parameter:
+                      1<=p<=infinity
+                      2: Euclidean distance
+                      1: Manhattan distance
+                      Ignored if the KDTree is an ArcKDTree
+        """
+        #first, coerce to columnar point array
+        new_data = np.asarray(new_data).reshape(-1,2)
+        if new_data is not None:
+            # If there's new data, we have to refit a tree
+            data = np.vstack((self.data, new_data))
+            if new_ids is not None:
+                ids = self.id_order.extend(list(new_ids))
+            else:
+                ids = range(data.shape[0])
+        elif (new_data is None) and (new_ids is None):
+            # If not, we can use the same kdtree we have
+            data = self.kdtree
+            ids = self.id_order
+        elif (new_data is None) and (new_ids is not None):
+            Warn('Remapping ids must be done using w.remap_ids')
+        if k is None:
+            k = self.k
+        if p is None:
+            p = self.p
+        if inplace:
+            self._reset()
+            self.__init__(data, ids=ids, k=k, p=p)
+        else:
+            return KNN(data, ids=ids, k=k, p=p)
 
 class Kernel(W):
     """
@@ -295,6 +486,69 @@ class Kernel(W):
             for i in neighbors:
                 weights[i][neighbors[i].index(i)] = 1.0
         W.__init__(self, neighbors, weights, ids)
+    
+    @classmethod
+    def from_shapefile(cls, filepath, idVariable=None,  **kwargs):
+        """
+        Kernel based weights from shapefile
+
+        Arguments
+        ---------
+        shapefile   : string
+                      shapefile name with shp suffix
+        idVariable  : string
+                      name of column in shapefile's DBF to use for ids
+
+        Returns
+        --------
+        Kernel Weights Object
+
+        See Also
+        ---------
+        Kernel
+        """
+        points = get_points_array_from_shapefile(filepath)
+        if idVariable is not None:
+            ids = get_ids(filepath, idVariable)
+        else:
+            ids = None
+        return cls.from_array(points, ids=ids, **kwargs)
+    
+    @classmethod
+    def from_array(cls, array, **kwargs):
+        """
+        Right now, just an alias for Kernel(array,...) because the current
+        constructor can work on arrays and kdtrees
+        """
+        return cls(array, **kwargs)
+
+    @classmethod
+    def from_dataframe(cls, df, geom_col='geometry', ids=None, **kwargs):
+        """
+        Make Kernel weights from a dataframe.
+
+        Parameters
+        ----------
+        df      :   pandas.dataframe
+                    a dataframe with a geometry column that can be used to
+                    construct a W object
+        geom_col :   string
+                    column name of the geometry stored in df
+        ids     :   string or iterable
+                    if string, the column name of the indices from the dataframe
+                    if iterable, a list of ids to use for the W
+                    if None, df.index is used.
+
+        See Also
+        --------
+        Kernel
+        """
+        pts = get_points_array(df[geom_col])
+        if ids is None:
+            ids = df.index.tolist()
+        elif isinstance(ids, str):
+            ids = df[ids].tolist()
+        return cls(pts, ids=ids, **kwargs)
 
     def _k_to_W(self, ids=None):
         allneighbors = {}
@@ -473,6 +727,70 @@ class DistanceBand(W):
         self._band()
         neighbors, weights = self._distance_to_W(ids)
         W.__init__(self, neighbors, weights, ids)
+
+    @classmethod
+    def from_shapefile(cls, filepath, threshold, idVariable=None, **kwargs):
+        """
+        Distance-band based weights from shapefile
+
+        Arguments
+        ---------
+        shapefile   : string
+                      shapefile name with shp suffix
+        idVariable  : string
+                      name of column in shapefile's DBF to use for ids
+
+        Returns
+        --------
+        Kernel Weights Object
+
+        See Also
+        ---------
+        Kernel
+        """
+        points = get_points_array_from_shapefile(filepath)
+        if idVariable is not None:
+            ids = get_ids(filepath, idVariable)
+        else:
+            ids = None
+        return cls.from_array(points, threshold, ids=ids, **kwargs)
+    
+    @classmethod
+    def from_array(cls, array, threshold, **kwargs):
+        """
+        Right now, just an alias for DistanceBand(array,...)
+        because the current
+        constructor can work on arrays and kdtrees
+        """
+        return cls(array, threshold, **kwargs)
+    
+    @classmethod
+    def from_dataframe(cls, df, threshold, geom_col='geometry', ids=None, **kwargs):
+        """
+        Make DistanceBand weights from a dataframe.
+
+        Parameters
+        ----------
+        df      :   pandas.dataframe
+                    a dataframe with a geometry column that can be used to
+                    construct a W object
+        geom_col :   string
+                    column name of the geometry stored in df
+        ids     :   string or iterable
+                    if string, the column name of the indices from the dataframe
+                    if iterable, a list of ids to use for the W
+                    if None, df.index is used.
+
+        See Also
+        --------
+        DistanceBand
+        """
+        pts = get_points_array(df[geom_col])
+        if ids is None:
+            ids = df.index.tolist()
+        elif isinstance(ids, str):
+            ids = df[ids].tolist()
+        return cls(pts, threshold, ids=ids, **kwargs)
 
     def _band(self):
         """Find all pairs within threshold.
