+++ pysal/weights/tests/old/test_Wsets.py
