+++ pysal/weights/tests/old/test__contW_binning.py
