+++ pysal/weights/tests/old/test__contW_rtree.py
