+++ pysal/weights/tests/old/test_spatial_lag.py
