+++ pysal/weights/tests/old/test_user.py
