+++ pysal/weights/tests/test_Distance.py
@@ -1,161 +1,280 @@
-import os
-import unittest
-import pysal
+from pysal.weights2 import Distance as d
+from pysal.weights2.util import get_points_array
+from pysal.weights2 import Contiguity as c
+from pysal.common import RTOL, ATOL
+from pysal.cg.kdtree import KDTree
 import numpy as np
-from itertools import product
+import pysal as ps
+import unittest as ut
 
+PANDAS_EXTINCT = ps.common.pandas is None
+# All instances should test these four methods, and define their own functional
+# tests based on common codepaths/estimated weights use cases. 
 
-class TestDistanceWeights(unittest.TestCase):
+class Any_Distance(ut.TestCase):
     def setUp(self):
-        np.random.seed(1234)
-        self.polyShp = pysal.examples.get_path('columbus.shp')
-        self.arcShp = pysal.examples.get_path('stl_hom.shp')
-        self.points = [(
-            10, 10), (20, 10), (40, 10), (15, 20), (30, 20), (30, 30)]
+        self.polygon_path = ps.examples.get_path('columbus.shp')
+        self.arc_path = ps.examples.get_path('stl_hom.shp')
+        self.points = [(10, 10), (20, 10), (40, 10), 
+                       (15, 20), (30, 20), (30, 30)]
+        self.euclidean_kdt = ps.cg.KDTree(self.points, distance_metric='euclidean')
+        
+        self.polygon_f = ps.open(self.polygon_path) # our file handler
+        self.poly_centroids = get_points_array(self.polygon_f) # our iterable
+        self.polygon_f.seek(0) #go back to head of file
+        
+        self.arc_f = ps.open(self.arc_path)
+        arc = ps.cg.sphere.arcdist
+        self.arc_points = get_points_array(self.arc_f)
+        self.arc_f.seek(0)
+        self.arc_kdt = ps.cg.KDTree(self.arc_points, distance_metric='Arc',
+                                    radius=ps.cg.sphere.RADIUS_EARTH_KM)
+        
+        # Without requiring users to have GDAL to run tests, 
+        # we can mock a shapely object by an object with a geo_interface
+        self.cls = object # class constructor
+        self.known_wi = None #index of known w entry to compare
+        self.known_w = dict() #actual w entry
+        self.known_name = self.known_wi
+            
+    def test_init(self):
+        # test vanilla, named
+        raise NotImplementedError('You need to implement this test '
+                                  'before this module will pass')
 
-    def test_knnW(self):
-        kd = pysal.cg.kdtree.KDTree(np.array(self.points), distance_metric='euclidean')
-        wnn2 = pysal.knnW(kd, 2)
-        self.assertEqual(wnn2.neighbors[0], [1,3])
+    def test_from_shapefile(self):
+        # test vanilla, named, sparse
+        raise NotImplementedError('You need to implement this test '
+                                  'before this module will pass')
 
-        pts = [i.centroid for i in pysal.open(self.polyShp)]
-        kd = pysal.cg.kdtree.KDTree(pts)
-        wnn4 = pysal.knnW(kd, 4)
-        self.assertEqual(set(wnn4.neighbors[0]), set([2,1,3,7]))
-        self.assertEqual(set(wnn4.neighbors[7]), set([3,6,12,11]))
+    def test_from_array(self):
+        # test named, sparse
+        raise NotImplementedError('You need to implement this test '
+                                  'before this module will pass')
 
-    def test_knnW_arc(self):
-        pts = [x.centroid for x in pysal.open(self.arcShp)]
-        dist = pysal.cg.sphere.arcdist  # default radius is Earth KM
-        full = np.matrix([[dist(pts[i], pts[j]) for j in xrange(
-            len(pts))] for i in xrange(len(pts))])
+    def test_from_dataframe(self):
+        # test named, columnar, defau
+        raise NotImplementedError('You need to implement this test '
+                                  'before this module will pass')
 
-        kd = pysal.cg.kdtree.KDTree(pts, distance_metric='Arc',
-                                    radius=pysal.cg.sphere.RADIUS_EARTH_KM)
-        w = pysal.knnW(kd, 4)
-        self.assertEqual(set(w.neighbors[4]), set([1, 3, 9, 12]))
-        self.assertEqual(set(w.neighbors[40]), set([31, 38, 45, 49]))
-        # self.assertTrue((full.argsort()[:, 1:5] == np.array(
-        #    [w.neighbors[x] for x in range(len(pts))])).all())
+class Test_KNN(Any_Distance):
+    def setUp(self):
+        Any_Distance.setUp(self)
+        
+        self.known_wi0 = 7
+        self.known_w0 = [3, 6, 12, 11]
+        self.known_wi1 = 0
+        self.known_w1 = [2, 1, 3 ,7]
 
-    def test_Kernel(self):
-        kw = pysal.Kernel(self.points)
-        wds = {kw.neighbors[0][i]: v for i, v in enumerate(kw.weights[0])}
-        self.assertEqual(wds, {0: 1, 1: 0.500000049999995, 3: 0.4409830615267465})
-        kw15 = pysal.Kernel(self.points, bandwidth=15.0)
-        self.assertEqual(kw15[0], {0: 1.0, 1: 0.33333333333333337,
-                                   3: 0.2546440075000701})
-        self.assertEqual(kw15.bandwidth[0], 15.)
-        self.assertEqual(kw15.bandwidth[-1], 15.)
-        bw = [25.0, 15.0, 25.0, 16.0, 14.5, 25.0]
-        kwa = pysal.Kernel(self.points, bandwidth=bw)
-        wds = {kwa.neighbors[0][i]: v for i, v in enumerate(kwa.weights[0])}
-        self.assertEqual(wds, {0: 1.0, 1: 0.59999999999999998,
-                               3: 0.55278640450004202, 4: 0.10557280900008403})
-        self.assertEqual(kwa.bandwidth[0], 25.)
-        self.assertEqual(kwa.bandwidth[1], 15.)
-        self.assertEqual(kwa.bandwidth[2], 25.)
-        self.assertEqual(kwa.bandwidth[3], 16.)
-        self.assertEqual(kwa.bandwidth[4], 14.5)
-        self.assertEqual(kwa.bandwidth[5], 25.)
-        kwea = pysal.Kernel(self.points, fixed=False)
-        wds = {kwea.neighbors[0][i]: v for i, v in enumerate(kw.weights[0])}
-        self.assertEqual(kwea.weights[0], [1.0, 0.10557289844279438,
-                                           9.9999990066379496e-08])
-        l = kwea.bandwidth.tolist()
-        np.allclose(l, [[11.180341005532938], [11.180341005532938],
-                        [20.000002000000002], [11.180341005532938],
-                        [14.142137037944515], [18.027758180095585]])
-        kweag = pysal.Kernel(self.points, fixed=False, function='gaussian')
-        wds = {kweag.neighbors[0][i]: v for i, v in enumerate(kweag.weights[0])}
-        self.assertEqual(wds, {0: 0.3989422804014327,
-                               1: 0.26741902915776961,
-                               3: 0.24197074871621341})
-        l = kweag.bandwidth.tolist()
-        np.allclose(l, [[11.180341005532938], [11.180341005532938],
-                        [20.000002000000002], [11.180341005532938],
-                        [14.142137037944515], [18.027758180095585]])
+        self.known_wi2 = 4
+        self.known_w2 = [1, 3, 9, 12]
+        self.known_wi3 = 40
+        self.known_w3 = [31, 38, 45, 49]
+    
+    ##########################
+    # Classmethod tests      #
+    ##########################
 
-        kw = pysal.kernelW_from_shapefile(self.polyShp, idVariable='POLYID')
-        wds = {kw.neighbors[1][i]: v for i, v in enumerate(kw.weights[1])}
-        self.assertEqual(wds, {4: 0.0070787731484506233,
-                               2: 0.2052478782400463,
-                               3: 0.23051223027663237,
-                               1: 1.0})
-        kwa = pysal.adaptive_kernelW_from_shapefile(self.polyShp)
-        wds = {kwa.neighbors[0][i]: v for i, v in enumerate(kwa.weights[0])}
-        self.assertEqual(wds, {0: 1.0, 2: 0.03178906767736345,
-                               1: 9.9999990066379496e-08})
+    def test_init(self):
+        w = d.KNN(self.euclidean_kdt, k=2)
+        self.assertEqual(w.neighbors[0], [1,3])
 
-    def test_threshold(self):
-        md = pysal.min_threshold_dist_from_shapefile(self.polyShp)
-        self.assertEqual(md, 0.61886415807685413)
-        wid = pysal.threshold_continuousW_from_array(self.points, 11.2)
-        wds = {wid.neighbors[0][i]: v for i, v in enumerate(wid.weights[0])}
-        self.assertEqual(wds, {1: 0.10000000000000001,
-                               3: 0.089442719099991588})
-        wid2 = pysal.threshold_continuousW_from_array(
-            self.points, 11.2, alpha=-2.0)
-        wds = {wid2.neighbors[0][i]: v for i, v in enumerate(wid2.weights[0])}
-        self.assertEqual(wds, {1: 0.01, 3: 0.0079999999999999984})
-        w = pysal.threshold_continuousW_from_shapefile(
-            self.polyShp, 0.62, idVariable="POLYID")
-        wds = {w.neighbors[1][i]: v for i, v in enumerate(w.weights[1])}
-        self.assertEqual(wds, {2: 1.6702346893743334,
-                               3: 1.7250729841938093})
+    @ut.skipIf(PANDAS_EXTINCT, 'Missing pandas')
+    def test_from_dataframe(self):
+        df = ps.pdio.read_files(self.polygon_path)
+        w = d.KNN.from_dataframe(df, k=4)
+        self.assertEqual(w.neighbors[self.known_wi0], self.known_w0)
+        self.assertEqual(w.neighbors[self.known_wi1], self.known_w1)
 
-    def test_DistanceBand(self):
-        """ see issue #126 """
-        w = pysal.rook_from_shapefile(
-            pysal.examples.get_path("lattice10x10.shp"))
-        polygons = pysal.open(
-            pysal.examples.get_path("lattice10x10.shp"), "r").read()
-        points1 = [poly.centroid for poly in polygons]
-        w1 = pysal.DistanceBand(points1, 1)
-        for k in range(w.n):
-            self.assertEqual(w[k], w1[k])
+    def test_from_array(self):
+        w = d.KNN.from_array(self.poly_centroids, k=4)
+        self.assertEqual(w.neighbors[self.known_wi0], self.known_w0)
+        self.assertEqual(w.neighbors[self.known_wi1], self.known_w1)
 
-    def test_DistanceBand_named(self):
-        w = pysal.rook_from_shapefile(
-                pysal.examples.get_path('lattice10x10.shp'))
-        polygons = pysal.open(
-                pysal.examples.get_path('lattice10x10.shp'),'r').read()
-        points1 = [poly.centroid for poly in polygons]
-        # two-letter names: 'AA', 'AB', .. 'JJ'
-        names = [''.join((chr(65+i), chr(65+j))) for i,j in product(range(10), range(10))]
-        w1 = pysal.DistanceBand(points1, 1, ids=names)
-        for k in range(w.n):
-            #need heaver comparison because looking between named & unnamed
-            name = names[k]
-            self.assertEqual(set(w.weights[k]), set(w1.weights[name]))
-            idxs_as_names = [names[i] for i in w.neighbors[k]]
-            self.assertEqual(set(idxs_as_names), set(w1.neighbors[name]))
+    def test_from_shapefile(self):
+        w = d.KNN.from_shapefile(self.polygon_path, k=4)    
+        self.assertEqual(w.neighbors[self.known_wi0], self.known_w0)
+        self.assertEqual(w.neighbors[self.known_wi1], self.known_w1)
 
-    def test_DistanceBand_ints(self):
-        """ see issue #126 """
-        w = pysal.rook_from_shapefile(
-            pysal.examples.get_path("lattice10x10.shp"))
-        polygons = pysal.open(
-            pysal.examples.get_path("lattice10x10.shp"), "r").read()
-        points2 = [tuple(map(int, poly.vertices[0])) for poly in polygons]
-        w2 = pysal.DistanceBand(points2, 1)
-        for k in range(w.n):
-            self.assertEqual(w[k], w2[k])
+    ##########################
+    # Function/User tests    #
+    ##########################
 
-    def test_DistanceBand_arc(self):
-        pts = [x.centroid for x in pysal.open(self.arcShp)]
-        dist = pysal.cg.sphere.arcdist  # default radius is Earth KM
-        full = np.matrix([[dist(pts[i], pts[j]) for j in xrange(
-            len(pts))] for i in xrange(len(pts))])
+    def test_reweight(self):
+        w = d.KNN(self.points, k=2)
+        new_point = [(21,21)]
+        wnew = w.reweight(k=4, p=1, new_data=new_point, inplace=False)
+        self.assertEqual(wnew[0], {1: 1.0, 3: 1.0, 4: 1.0, 6: 1.0})
 
-        kd = pysal.cg.kdtree.KDTree(pts, distance_metric='Arc',
-                                    radius=pysal.cg.sphere.RADIUS_EARTH_KM)
-        w = pysal.DistanceBand(kd, full.max(), binary=False, alpha=1.0)
-        self.assertTrue((w.sparse.todense() == full).all())
+class Test_DistanceBand(Any_Distance):
+    def setUp(self):
+        Any_Distance.setUp(self)
+        self.grid_path =  ps.examples.get_path('lattice10x10.shp')
+        self.grid_rook_w = c.Rook.from_shapefile(self.grid_path)
+        self.grid_f = ps.open(self.grid_path)
+        self.grid_points = get_points_array(self.grid_f)
+        self.grid_f.seek(0)
 
+        self.grid_kdt = KDTree(self.grid_points)
+    
+    ##########################
+    # Classmethod tests      #
+    ##########################
 
-suite = unittest.TestLoader().loadTestsFromTestCase(TestDistanceWeights)
+    def test_init(self):
+        w = d.DistanceBand(self.grid_kdt, 1)
+        for k,v in w:
+            self.assertEquals(v, self.grid_rook_w[k])
 
+    def test_from_shapefile(self):
+        w = d.DistanceBand.from_shapefile(self.grid_path, 1)
+        for k,v in w:
+            self.assertEquals(v, self.grid_rook_w[k])
+
+    def test_from_array(self):
+        w = d.DistanceBand.from_array(self.grid_points, 1)
+        for k,v in w:
+            self.assertEquals(v, self.grid_rook_w[k])
+
+    @ut.skipIf(PANDAS_EXTINCT, 'Missing pandas')
+    def test_from_dataframe(self):
+        import pandas as pd
+        geom_series = ps.pdio.shp.shp2series(self.grid_path)
+        random_data = np.random.random(size=len(geom_series))
+        df = pd.DataFrame({'obs':random_data, 'geometry':geom_series})
+        w = d.DistanceBand.from_dataframe(df, 1)
+        for k,v in w:
+            self.assertEquals(v, self.grid_rook_w[k])
+
+    ##########################
+    # Function/User tests    #
+    ##########################
+    def test_integers(self):
+        """
+        see issue #126
+        """
+        grid_integers = [tuple(map(int, poly.vertices[0])) 
+                              for poly in self.grid_f]
+        self.grid_f.seek(0)
+        grid_dbw = d.DistanceBand(grid_integers, 1)
+        for k,v in grid_dbw:
+            self.assertEquals(v, self.grid_rook_w[k])
+
+    def test_arcdist(self):
+        arc = ps.cg.sphere.arcdist
+        kdt = KDTree(self.arc_points, distance_metric='Arc',
+                     radius=ps.cg.sphere.RADIUS_EARTH_KM)
+        npoints = self.arc_points.shape[0]
+        full = np.matrix([[arc(self.arc_points[i], self.arc_points[j])
+                          for j in xrange(npoints)] 
+                          for i in xrange(npoints)])
+        maxdist = full.max()
+        w = d.DistanceBand(kdt, maxdist, binary=False, alpha=1.0)
+        np.testing.assert_allclose(w.sparse.todense(), full)
+
+class Test_Kernel(Any_Distance):
+    def setUp(self):
+        Any_Distance.setUp(self)
+
+        self.known_wi0 = 0
+        self.known_w0 = {0: 1, 1: 0.500000049999995, 3: 0.4409830615267465}
+
+        self.known_wi1 = 0
+        self.known_w1 = {0: 1.0, 1: 0.33333333333333337,
+                         3: 0.2546440075000701}
+        self.known_w1_bw = 15.
+
+        self.known_wi2 = 0
+        self.known_w2 = {0: 1.0, 1: 0.59999999999999998,
+                         3: 0.55278640450004202, 4: 0.10557280900008403}
+        self.known_w2_bws = [25.0, 15.0, 25.0, 16.0, 14.5, 25.0]
+
+        self.known_wi3 = 0
+        self.known_w3 = [1.0, 0.10557289844279438, 9.9999990066379496e-08]
+        self.known_w3_abws =[[11.180341005532938], [11.180341005532938],
+                             [20.000002000000002], [11.180341005532938],
+                             [14.142137037944515], [18.027758180095585]]
+
+        self.known_wi4 = 0
+        self.known_w4 = {0: 0.3989422804014327,
+                         1: 0.26741902915776961,
+                         3: 0.24197074871621341}
+        self.known_w4_abws = self.known_w3_abws
+
+        self.known_wi5 = 1
+        self.known_w5 = {4: 0.0070787731484506233,
+                         2: 0.2052478782400463,
+                         3: 0.23051223027663237,
+                         1: 1.0}
+
+        self.known_wi6 = 0
+        self.known_w6 = {0: 1.0, 2: 0.03178906767736345,
+                         1: 9.9999990066379496e-08}
+        #stick answers & params here
+
+    ##########################
+    # Classmethod tests      #
+    ##########################
+
+    def test_init(self):
+        w = d.Kernel(self.euclidean_kdt)
+        for k,v in w[self.known_wi0].items():
+            np.testing.assert_allclose(v, self.known_w0[k], rtol=RTOL)
+
+    def test_from_shapefile(self):
+        w = d.Kernel.from_shapefile(self.polygon_path, idVariable='POLYID')
+        for k,v in w[self.known_wi5].items():
+            np.testing.assert_allclose((k,v), (k,self.known_w5[k]), rtol=RTOL)
+        
+        w = d.Kernel.from_shapefile(self.polygon_path, fixed=False)
+        for k,v in w[self.known_wi6].items():
+            np.testing.assert_allclose((k,v), (k,self.known_w6[k]), rtol=RTOL)
+
+    def test_from_array(self):
+        w = d.Kernel.from_array(self.points)
+        for k,v in w[self.known_wi0].items():
+            np.testing.assert_allclose(v, self.known_w0[k], rtol=RTOL)
+    
+    @ut.skipIf(PANDAS_EXTINCT, 'Missing pandas')
+    def test_from_dataframe(self):
+        df = ps.pdio.read_files(self.polygon_path)
+        w = d.Kernel.from_dataframe(df)
+        for k,v in w[self.known_wi5-1].items():
+            np.testing.assert_allclose(v, self.known_w5[k+1], rtol=RTOL)
+    
+    ##########################
+    # Function/User tests    # 
+    ##########################
+
+    def test_fixed_bandwidth(self):
+        w = d.Kernel(self.points, bandwidth=15.0)
+        for k,v in w[self.known_wi1].items():
+            np.testing.assert_allclose((k,v), (k, self.known_w1[k]))
+        np.testing.assert_allclose(np.ones((w.n,1))*15, w.bandwidth)
+
+        w = d.Kernel(self.points, bandwidth=self.known_w2_bws)
+        for k,v in w[self.known_wi2].items():
+            np.testing.assert_allclose((k,v), (k, self.known_w2[k]), rtol=RTOL)
+        for i in range(w.n):
+            np.testing.assert_allclose(w.bandwidth[i], self.known_w2_bws[i], rtol=RTOL)
+    
+    def test_adaptive_bandwidth(self):
+        w = d.Kernel(self.points, fixed=False)
+        np.testing.assert_allclose(sorted(w[self.known_wi3].values()),
+                                   sorted(self.known_w3), rtol=RTOL)
+        bws = w.bandwidth.tolist()
+        np.testing.assert_allclose(bws, self.known_w3_abws, rtol=RTOL)
+
+        w = d.Kernel(self.points, fixed=False, function='gaussian')
+        for k,v in w[self.known_wi4].items():
+            np.testing.assert_allclose((k,v), (k, self.known_w4[k]), rtol=RTOL)
+        bws = w.bandwidth.tolist()
+        np.testing.assert_allclose(bws, self.known_w4_abws, rtol=RTOL)
+
+knn = ut.TestLoader().loadTestsFromTestCase(Test_KNN)
+kern = ut.TestLoader().loadTestsFromTestCase(Test_Kernel)
+db = ut.TestLoader().loadTestsFromTestCase(Test_DistanceBand)
+suite = ut.TestSuite([knn, kern, db])
 if __name__ == '__main__':
-    runner = unittest.TextTestRunner()
+    runner = ut.TextTestRunner()
     runner.run(suite)
