--- pysal/weights2/__init__.py
@@ -1,2 +1,0 @@
-from .Contiguity import Rook, Queen
-from .Distance import Kernel, DistanceBand, KNN
